# pokey_device::step_pot: remove operations with no effect