--- a/src/devices/sound/pokey.cpp
+++ b/src/devices/sound/pokey.cpp
@@ -538,11 +538,8 @@ void pokey_device::step_keyboard()
 
 void pokey_device::step_pot()
 {
-	if ((m_SKCTL & SK_RESET) == 0)
-		return;
-
-	uint8_t upd = 0;
 	m_pot_counter++;
+	uint8_t upd = 0;
 	for (int pot = 0; pot < 8; pot++)
 	{
 		if ((m_POTx[pot]<m_pot_counter) || (m_pot_counter == 228))
@@ -551,7 +548,9 @@ void pokey_device::step_pot()
 			/* latching is emulated in read */
 		}
 	}
-	synchronize(SYNC_POT, upd);
+	// some pots latched?
+	if (upd != 0)
+		synchronize(SYNC_POT, upd);
 }
 
 /*