* step_pot is called from step_one_clock only and just in case pokey is not in
  reset state -> No need to check reset state again
* in case there were no bits in 'upd' latched to one, there is no need to call
  synchronize(SYNC_POT, 0) because m_ALLPOT won't change.

Performance results with missile / starwars (no pots):
Before:
./mame64 -bench 50 missile -> Average speed: 1171.67% (49 seconds)
./mame64 -bench 50 starwars -> Average speed: 551.66% (49 seconds)
After:
./mame64 -bench 50 missile -> Average speed: 1321.16% (49 seconds)
./mame64 -bench 50 starwars -> Average speed: 551.10% (49 seconds)

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>